backend/pyproject.toml
[tool.poetry]
name = "bk_simulation_streaming_test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
numpy = "^1.22.3"
i = "^2.1.0"
websocket-client = "^1.3.2"
nest-asyncio = "^1.5.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
